refactor: 불필요 구문 제거
