+++ src/test/java/com/flab/just_10_minutes/Point/service/PointServiceTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.ArrayList;
 import java.util.Optional;
@@ -37,7 +36,6 @@ class PointServiceTest {
 
     @BeforeEach
     public void setUp() {
-        MockitoAnnotations.openMocks(this);
         target = new PointService(userDao, pointDao);
     }
